server/requirements.txt
pymongo==4.11.3
python-dotenv==1.0.1
flask==3.1.0